chore: Move track relations functions over
