+++ mobile/src/api/track.ts
@@ -9,24 +9,6 @@ import { iAsc, iDesc } from "~/lib/drizzle";
 import type { ScreenSortOptions } from "~/stores/ViewPreference/constants";
 
 //#region GET Methods
-/** Get the genres that this track has. */
-export async function getTrackGenres(id: string) {
-  const allTrackGenres = await db.query.tracksToGenres.findMany({
-    where: (fields, { eq }) => eq(fields.trackId, id),
-    columns: { genreName: true },
-  });
-  return allTrackGenres.map(({ genreName }) => genreName);
-}
-
-/** Get the names of the playlists that this track is in. */
-export async function getTrackPlaylists(id: string) {
-  const allTrackPlaylists = await db.query.tracksToPlaylists.findMany({
-    where: (fields, { eq }) => eq(fields.trackId, id),
-    columns: { playlistName: true },
-  });
-  return allTrackPlaylists.map(({ playlistName }) => playlistName);
-}
-
 export type SortedTracksMode = "sortedIds" | "sortedTracks";
 
 export type PreSortedTrack = {
