+++ mobile/src/navigation/screens/playlists/ModifyView.tsx
@@ -12,8 +12,8 @@ import {
 
 import { PagePlaceholder } from "~/navigation/components/Placeholder";
 import {
-  formatTrackForForm,
   ModifyPlaylistBase,
+  formatTrackForForm,
 } from "./components/ModifyViewBase";
 
 import { mutateGuardAsync } from "~/lib/react-query";
