+++ mobile/src/queries/keyStore.ts
@@ -1,14 +1,9 @@
 import { createQueryKeyStore } from "@lukemorales/query-key-factory";
 import { ne } from "drizzle-orm";
 
-import { db } from "~/db";
 import { playlists } from "~/db/schema";
 
-import {
-  getSortedTracks,
-  getTrackGenres,
-  getTrackPlaylists,
-} from "~/api/track";
+import { getSortedTracks } from "~/api/track";
 import { getAlbum, getAlbumsSummary } from "~/data/album/api";
 import { getArtist, getArtistsSummary } from "~/data/artist/api";
 import { getFavoriteLists } from "~/data/favorite/api";
@@ -17,7 +12,13 @@ import { getGenre, getGenresSummary } from "~/data/genre/api";
 import { getLyric, getLyricsSummary } from "~/data/lyric/api";
 import { getPlaylist, getPlaylistsSummary } from "~/data/playlist/api";
 import { getRecentMedia } from "~/data/recent/api";
-import { getTrack } from "~/data/track/api";
+import {
+  getTrack,
+  getTrackFavoriteStatus,
+  getTrackGenres,
+  getTrackLyrics,
+  getTrackPlaylists,
+} from "~/data/track/api";
 
 import { FavoritesPlaylistKey } from "~/modules/media/constants";
 import type { ScreenSortOptions } from "~/stores/ViewPreference/constants";
@@ -83,14 +84,7 @@ export const queries = createQueryKeyStore({
     }),
     forTrack: (trackId: string) => ({
       queryKey: [trackId],
-      // FIXME: Have a `getTracklyrics` in `~/data/track/api.ts`.
-      queryFn: async () => {
-        const data = await db.query.tracksToLyrics.findFirst({
-          where: (fields, { eq }) => eq(fields.trackId, trackId),
-          with: { lyric: true },
-        });
-        return data?.lyric ?? null;
-      },
+      queryFn: () => getTrackLyrics(trackId),
     }),
   },
   /** Query keys used in `useQuery` for playlists. */
@@ -116,17 +110,9 @@ export const queries = createQueryKeyStore({
       queryFn: () => getTrack(trackId),
       contextQueries: {
         isFavorite: {
+          // eslint-disable-next-line @tanstack/query/exhaustive-deps
           queryKey: null,
-          queryFn: async () => {
-            const isFavorited = await db.query.tracksToPlaylists.findFirst({
-              where: (fields, { and, eq }) =>
-                and(
-                  eq(fields.playlistName, FavoritesPlaylistKey),
-                  eq(fields.trackId, trackId),
-                ),
-            });
-            return isFavorited ? true : false;
-          },
+          queryFn: () => getTrackFavoriteStatus(trackId),
         },
         genres: {
           // eslint-disable-next-line @tanstack/query/exhaustive-deps
